+++ include/fiction/technology/cell_technologies.hpp
@@ -18,7 +18,7 @@ struct qca_technology
     /**
      * Possible types of QCA cells.
      */
-    enum cell_type : uint8_t  // NOLINT(cppcoreguidelines-use-enum-class)
+    enum class cell_type : uint8_t
     {
         /**
          * Symbol used for empty QCA cells.
@@ -60,39 +60,39 @@ struct qca_technology
      */
     enum class cell_mark : uint8_t
     {
-        EMPTY  = cell_type::EMPTY,
-        INPUT  = cell_type::INPUT,
-        OUTPUT = cell_type::OUTPUT
+        EMPTY  = static_cast<uint8_t>(cell_type::EMPTY),
+        INPUT  = static_cast<uint8_t>(cell_type::INPUT),
+        OUTPUT = static_cast<uint8_t>(cell_type::OUTPUT)
     };
 
     [[nodiscard]] static constexpr bool is_empty_cell(const cell_type& c) noexcept
     {
-        return c == EMPTY;
+        return c == cell_type::EMPTY;
     }
 
     [[nodiscard]] static constexpr bool is_normal_cell(const cell_type& c) noexcept
     {
-        return c == NORMAL;
+        return c == cell_type::NORMAL;
     }
 
     [[nodiscard]] static constexpr bool is_input_cell(const cell_type& c) noexcept
     {
-        return c == INPUT;
+        return c == cell_type::INPUT;
     }
 
     [[nodiscard]] static constexpr bool is_output_cell(const cell_type& c) noexcept
     {
-        return c == OUTPUT;
+        return c == cell_type::OUTPUT;
     }
 
     [[nodiscard]] static constexpr bool is_const_0_cell(const cell_type& c) noexcept
     {
-        return c == CONST_0;
+        return c == cell_type::CONST_0;
     }
 
     [[nodiscard]] static constexpr bool is_const_1_cell(const cell_type& c) noexcept
     {
-        return c == CONST_1;
+        return c == cell_type::CONST_1;
     }
 
     [[nodiscard]] static constexpr bool is_constant_cell(const cell_type& c) noexcept
@@ -343,7 +343,7 @@ struct inml_technology
     /**
      * Possible types of iNML cells.
      */
-    enum cell_type : uint8_t  // NOLINT(cppcoreguidelines-use-enum-class)
+    enum class cell_type : uint8_t
     {
         /**
          * Symbol used for empty iNML cells.
@@ -393,39 +393,39 @@ struct inml_technology
      */
     enum class cell_mark : uint8_t
     {
-        EMPTY  = cell_type::EMPTY,
-        INPUT  = cell_type::INPUT,
-        OUTPUT = cell_type::OUTPUT
+        EMPTY  = static_cast<uint8_t>(cell_type::EMPTY),
+        INPUT  = static_cast<uint8_t>(cell_type::INPUT),
+        OUTPUT = static_cast<uint8_t>(cell_type::OUTPUT)
     };
 
     [[nodiscard]] static constexpr bool is_empty_cell(const cell_type& c) noexcept
     {
-        return c == EMPTY;
+        return c == cell_type::EMPTY;
     }
 
     [[nodiscard]] static constexpr bool is_normal_cell(const cell_type& c) noexcept
     {
-        return c == NORMAL;
+        return c == cell_type::NORMAL;
     }
 
     [[nodiscard]] static constexpr bool is_input_cell(const cell_type& c) noexcept
     {
-        return c == INPUT;
+        return c == cell_type::INPUT;
     }
 
     [[nodiscard]] static constexpr bool is_output_cell(const cell_type& c) noexcept
     {
-        return c == OUTPUT;
+        return c == cell_type::OUTPUT;
     }
 
     [[nodiscard]] static constexpr bool is_slanted_edge_up_magnet(const cell_type& c) noexcept
     {
-        return c == SLANTED_EDGE_UP_MAGNET;
+        return c == cell_type::SLANTED_EDGE_UP_MAGNET;
     }
 
     [[nodiscard]] static constexpr bool is_slanted_edge_down_magnet(const cell_type& c) noexcept
     {
-        return c == SLANTED_EDGE_DOWN_MAGNET;
+        return c == cell_type::SLANTED_EDGE_DOWN_MAGNET;
     }
 
     [[nodiscard]] static constexpr bool is_slanted_edge_magnet(const cell_type& c) noexcept
@@ -435,17 +435,17 @@ struct inml_technology
 
     [[nodiscard]] static constexpr bool is_inverter_magnet(const cell_type& c) noexcept
     {
-        return c == INVERTER_MAGNET;
+        return c == cell_type::INVERTER_MAGNET;
     }
 
     [[nodiscard]] static constexpr bool is_crosswire_magnet(const cell_type& c) noexcept
     {
-        return c == CROSSWIRE_MAGNET;
+        return c == cell_type::CROSSWIRE_MAGNET;
     }
 
     [[nodiscard]] static constexpr bool is_fanout_coupler_magnet(const cell_type& c) noexcept
     {
-        return c == FANOUT_COUPLER_MAGNET;
+        return c == cell_type::FANOUT_COUPLER_MAGNET;
     }
 
     [[nodiscard]] static constexpr bool is_normal_cell_mode([[maybe_unused]] const cell_mode& m) noexcept
@@ -482,7 +482,7 @@ struct sidb_technology
     /**
      * Possible types of SiDB cells.
      */
-    enum cell_type : uint8_t  // NOLINT(cppcoreguidelines-use-enum-class)
+    enum class cell_type : uint8_t
     {
         /**
          * Symbol used for empty SiDB cells.
@@ -516,35 +516,35 @@ struct sidb_technology
      */
     enum class cell_mark : uint8_t
     {
-        EMPTY  = cell_type::EMPTY,
-        INPUT  = cell_type::INPUT,
-        OUTPUT = cell_type::OUTPUT,
-        LOGIC  = cell_type::LOGIC
+        EMPTY  = static_cast<uint8_t>(cell_type::EMPTY),
+        INPUT  = static_cast<uint8_t>(cell_type::INPUT),
+        OUTPUT = static_cast<uint8_t>(cell_type::OUTPUT),
+        LOGIC  = static_cast<uint8_t>(cell_type::LOGIC)
     };
 
     [[nodiscard]] static constexpr bool is_empty_cell(const cell_type& c) noexcept
     {
-        return c == EMPTY;
+        return c == cell_type::EMPTY;
     }
 
     [[nodiscard]] static constexpr bool is_normal_cell(const cell_type& c) noexcept
     {
-        return c == NORMAL;
+        return c == cell_type::NORMAL;
     }
 
     [[nodiscard]] static constexpr bool is_input_cell(const cell_type& c) noexcept
     {
-        return c == INPUT;
+        return c == cell_type::INPUT;
     }
 
     [[nodiscard]] static constexpr bool is_output_cell(const cell_type& c) noexcept
     {
-        return c == OUTPUT;
+        return c == cell_type::OUTPUT;
     }
 
     [[nodiscard]] static constexpr bool is_logic_cell(const cell_type& c) noexcept
     {
-        return c == LOGIC;
+        return c == cell_type::LOGIC;
     }
 
     [[nodiscard]] static constexpr bool is_normal_cell_mode([[maybe_unused]] const cell_mode& m) noexcept